Update blog.md
I added the link to my image for the blog keyword and formatted the citations in Markdown language. However, I'm concerned about the citation style, but I haven't italicized either the webpage title or the image. Is it important for endnotes to be in alphabetical order? It seems challenging to achieve that in Markdown.